+++ lib/transpec/configuration.rb
@@ -16,6 +16,7 @@ module Transpec
       [:convert_deprecated_method,                             true],
       [:parenthesize_matcher_arg,                              true],
       [:add_receiver_arg_to_any_instance_implementation_block, true],
+      [:convert_stub_with_hash_to_stub_and_return,             false],
       [:forced,                                                false],
       [:skip_dynamic_analysis,                                 false]
     ].freeze